move json serializer to serializers.py
