+++ rq/serializers.py
@@ -1,5 +1,6 @@
 from functools import partial
 import pickle
+import json
 
 from .compat import string_types
 from .utils import import_attribute
@@ -8,6 +9,16 @@ from .utils import import_attribute
 class DefaultSerializer:
     dumps = partial(pickle.dumps, protocol=pickle.HIGHEST_PROTOCOL)
     loads = pickle.loads
+
+
+class JSONSerializer():
+    @staticmethod
+    def dumps(*args, **kwargs):
+        return json.dumps(*args, **kwargs).encode('utf-8')
+
+    @staticmethod
+    def loads(s, *args, **kwargs):
+        return json.loads(s.decode('utf-8'), *args, **kwargs)
 
 
 def resolve_serializer(serializer):
